I've managed to get the WindowSkin working correctly. now, we need to make a cleanup and some fixes on the Window, TailedWindow e SlimWindow classes. And the Kernel is leaving a lot of memory leaks, we must fix it, and later, we'll need to rewrite it.
All Window classes now have a little bug, if you don't resize the window, it will not show up,
we must take a look at that, also, it doesn't update when you change the window alpha, and in the SlimWindow, the VertexColor doesn't affect it.

And, the WindowSkin export and import functions seems to be working correctly too, it generated the file correctly and loaded it.
(The file itself is a bit small, less than 1kb, I think we should merge it with another file)

That remind me, we need to elaborate some "loading strategies", because the way Ogre handles file loading.

And yes, I'm using the commit message to leave you a message.
Well, I'm done for today, next time (that is, today, when I wake up) I'll try to fix the leaks and get the Window class working correctly.